move cd, exit, pwd to each file
