+++ srcs/builtin/builtin_pwd.c
@@ -6,8 +6,26 @@
 /*   By: ssoeno <ssoeno@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/27 15:56:59 by ssoeno            #+#    #+#             */
-/*   Updated: 2024/10/27 16:50:53 by ssoeno           ###   ########.fr       */
+/*   Updated: 2024/10/31 21:40:10 by ssoeno           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "minishell.h"
+#include "../includes/minishell.h"
+
+int	builtin_pwd(int argc, char *argv[])
+{
+	char	buf[PATH_MAX];
+
+	if (argc != 1)
+	{
+		printf("%s: wrong argument\n", argv[0]);
+		return (1);
+	}
+	if (!getcwd(buf, PATH_MAX))
+	{
+		printf("%s: cannot get working directory\n", argv[0]);
+		return (1);
+	}
+	printf("%s\n", buf);
+	return (0);
+}
